adding enrollee search facets
